feat:REQ-050426 cambios de campos en las clases para las listas

+++ src/main/java/com/svc/ventas/models/mapstruct/dto/BajoStockDTO.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -13,6 +15,6 @@ public class BajoStockDTO {
 
   private String producto;
 
-  private Integer stock;
+  private BigDecimal stock;
 
 }
